refactor: move klaro code to mixin
